1. revert the int-to-long changes 2. merge Avx512F, BW, CD, DQ to a converged ISA flag, Avx512. 3. introduce APX cpuid detection.

+++ src/coreclr/tools/aot/jitinterface/jitwrapper.cpp
@@ -53,7 +53,7 @@ DLL_EXPORT void JitProcessShutdownWork(ICorJitCompiler * pJit)
     return pJit->ProcessShutdownWork(nullptr);
 }
 
-DLL_EXPORT long long JitGetProcessorFeatures()
+DLL_EXPORT int JitGetProcessorFeatures()
 {
     return minipal_getcpufeatures();
 }
